# public/tools/schema/openapi_rest_demo.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST demo",
    "description": "This is a sample REST serverform jaktestowac.pl .",
    "termsOfService": "http://jaktestowac.pl/",
    "version": "1.2.0"
  },
...
  "servers": [
    {
      "url": "https://rest-demo-jt.herokuapp.com/api"
    }
  ],
...
  "paths": {
    "/articles": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Get articles",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "Id of article."
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "Id of user, creator of article."
          }
        ],
        "operationId": "getArticles",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articles"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Add article",
        "operationId": "postArticle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "Article created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - malformed JSON",
            "content": {}
          },
          "422": {
            "description": "Invalid article supplied",
            "content": {}
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get users",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create user",
        "operationId": "postUser",
        "requestBody": {
          "description": "Creats user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - malformed JSON",
            "content": {}
          },
          "409": {
            "description": "Conflict - Invalid user email",
            "content": {}
          },
          "422": {
            "description": "Invalid user supplied",
            "content": {}
          }
        }
      }
    },
...
    "/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "Get comments",
        "operationId": "getComments",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comments"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "basicAuth": []
          }
        ],
        "tags": [
          "comments"
        ],
        "summary": "Create comment",
        "operationId": "createComment",
        "requestBody": {
          "description": "Created Comment object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - malformed JSON",
            "content": {}
          },
          "422": {
            "description": "Invalid comment supplied",
            "content": {}
          }
        }
      }
    },
...
    "/articles?": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Get articles with querry",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Free search articles."
          },
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "Id of article."
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "Id of user, creator of article."
          },
          {
            "in": "query",
            "name": "_sort",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Sort with given attribute (user_id, title, id), use it with _order."
          },
          {
            "in": "query",
            "name": "_order",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Order default ascending (asc) for descending use desc."
          },
          {
            "in": "query",
            "name": "_limit",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "allowEmptyValue": true,
            "description": "Limit returned items. By default empty limit is 10."
          },
          {
            "in": "query",
            "name": "_page",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "Page number when using _limit."
          },
          {
            "in": "query",
            "name": "_start",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "How many items should be skipped."
          },
          {
            "in": "query",
            "name": "_end",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "How many to show after _start."
          },
          {
            "in": "query",
            "name": "_embed",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Embed depending objects (users, comments)."
          },
          {
            "in": "query",
            "name": "_expand",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Expand depending objects (users, comments)."
          }
        ],
        "operationId": "getArticlesQuery",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articles"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "Article": {
        "type": "object",
        "required": [
          "user_id",
          "title",
          "body",
          "date"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "Title"
          },
          "body": {
            "type": "string",
            "description": "Article body"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "image": {
            "type": "string",
            "description": "Image path"
          }
        },
        "xml": {
          "name": "Article"
        }
      },
      "Articles": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Article"
        }
      },
      "ArticleUpdate": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "Title"
          },
          "body": {
            "type": "string",
            "description": "Article body"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "image": {
            "type": "string",
            "description": "Image path"
          }
        },
        "xml": {
          "name": "Article"
        }
      },
      "Comment": {
        "type": "object",
        "required": [
          "user_id",
          "article_id",
          "body",
          "date"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "article_id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string",
            "description": "Comment body"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        },
        "xml": {
          "name": "Comment"
        }
      },
      "Comments": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Comment"
        }
      },
      "CommentUpdate": {
        "type": "object",
        "properties": {
          "article_id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string",
            "description": "Comment body"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        },
        "xml": {
          "name": "Comment"
        }
      },
      "Users": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/User"
        }
      },
      "User": {
        "type": "object",
        "required": [
          "email",
          "firstname",
          "lastname",
          "avatar"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User"
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User"
        }
      },
...
    }
  }
}